+++ server/impl/src/main/java/com/walmartlabs/concord/server/ServerResource.java
@@ -32,9 +32,6 @@ import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.Properties;
 
 @Named
 @Singleton