+++ server/impl/src/main/java/com/walmartlabs/concord/server/org/triggers/TriggerV2Resource.java
@@ -22,7 +22,6 @@ package com.walmartlabs.concord.server.org.triggers;
 
 import com.walmartlabs.concord.common.validation.ConcordKey;
 import com.walmartlabs.concord.server.org.OrganizationDao;
-import com.walmartlabs.concord.server.org.OrganizationManager;
 import com.walmartlabs.concord.server.org.ResourceAccessLevel;
 import com.walmartlabs.concord.server.org.project.ProjectAccessManager;
 import com.walmartlabs.concord.server.org.project.ProjectDao;